# business-central-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Business Central Backend API
  version: 0.1.0
  description: |
    Contract-first OpenAPI documentation for the Business Central backend.

    Health, authentication, authorization, commerce, reporting, service, and
    repair endpoints are implemented. Reusable schemas, security conventions,
    response envelopes, required error codes, and collection query semantics
    are documented here.
  license:
    name: Proprietary
  x-business-central-status: implemented-current-scope
servers:
  - url: http://localhost:8080
    description: Local backend (when the HTTP runtime is available)
...
paths:
...
  /api/v1/pos/catalog:
    get: { tags: [Catalog], summary: List sellable variants with current price and shop stock, operationId: listPOSCatalog, security: [{ bearerAuth: [] }], parameters: [{ name: shop_id, in: query, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: POS catalog, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/pos/quote:
    post: { tags: [Catalog], summary: Calculate an authoritative sale quote, operationId: quotePOSSale, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [lines], properties: { shop_id: { type: string, format: uuid }, customer_name: { type: string }, customer_phone: { type: string }, delivery_id: { type: string, format: uuid }, delivery_fee: { type: string }, manual_promotion: { type: string }, note: { type: string }, payment_method: { type: string }, promotion_id: { type: string, format: uuid }, lines: { type: array, items: { type: object, required: [variant_id, quantity], properties: { variant_id: { type: string, format: uuid }, quantity: { type: string } } } } } } } } }, responses: { '200': { description: "Server-calculated subtotal, discount, tax, delivery, and total", content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } } }
  /api/v1/pos/orders:
    post: { tags: [Catalog], summary: Complete an authoritative POS sale, operationId: createPOSSale, security: [{ bearerAuth: [] }], description: "Prices, promotion eligibility, customer creation, delivery fee, cash payment, fulfillment, inventory movements, audit, and ordered synchronization publication are committed atomically. The idempotency key prevents duplicate business effects. External payment methods are rejected until verified provider authorization is supplied through a dedicated provider flow.", requestBody: { required: true, content: { application/json: { schema: { type: object, required: [shop_id, lines, payment_method, idempotency_key], properties: { shop_id: { type: string, format: uuid }, customer_name: { type: string }, customer_phone: { type: string }, delivery_id: { type: string, format: uuid }, delivery_fee: { type: string }, manual_promotion: { type: string }, note: { type: string }, promotion_id: { type: string, format: uuid }, payment_method: { type: string, enum: [CASH] }, idempotency_key: { type: string }, lines: { type: array, items: { type: object, required: [variant_id, quantity], properties: { variant_id: { type: string, format: uuid }, quantity: { type: string } } } } } } } } }, responses: { '201': { description: Fulfilled order, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } }, '409': { description: External payment authorization required or idempotency conflict } } }
  /api/v1/pos/orders/{orderId}/refunds:
    post:
      tags: [Catalog]
      summary: Record a canonical POS payment refund
      operationId: createPOSRefund
      security: [{ bearerAuth: [] }]
      description: >-
        Records a successful refund against a captured payment after validating
        merchant and assigned-shop scope. The command is idempotent, cannot
        refund more than the payment's remaining captured amount, and updates
        payment/order status through the canonical database lifecycle triggers.
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_id, amount, idempotency_key]
              properties:
                payment_id: { type: string, format: uuid }
                amount: { type: string }
                reason: { type: string }
                refund_id: { type: string, format: uuid }
                idempotency_key: { type: string }
      responses:
        '201':
          description: Canonical refund created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefundEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/invoices:
    get: { tags: [Catalog], summary: List canonical sales and service invoices, operationId: listInvoices, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { description: "Invoices with printable line items, projection kind, repair ticket status/payment status, merchant, shop identity and optional shop logo/print preference, all service work items/devices, and printable ticket fields", content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
  /api/v1/inventory/locations:
    get: { tags: [Inventory], summary: List active tenant locations scoped to the staff shop, operationId: listInventoryLocations, security: [{ bearerAuth: [] }], responses: { '200': { description: Inventory locations, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } } }
...
  /api/v1/shops:
    get:
      tags: [Merchants]
      summary: List shops for the current merchant
      operationId: listShops
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: Merchant shops, content: { application/json: { schema: { $ref: '#/components/schemas/ShopCollectionEnvelope' } } } } }
    post:
      tags: [Merchants]
      summary: Create a shop for the current merchant
      operationId: createShop
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShopRequest' } } } }
      responses: { '201': { description: Shop created, content: { application/json: { schema: { $ref: '#/components/schemas/ShopEnvelope' } } } }, '409': { $ref: '#/components/responses/Conflict' } }
  /api/v1/shops/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Merchants]
      summary: Get a merchant shop
      operationId: getShop
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Shop, content: { application/json: { schema: { $ref: '#/components/schemas/ShopEnvelope' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    patch:
      tags: [Merchants]
      summary: Update a merchant shop
      operationId: updateShop
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ShopRequest' } } } }
      responses: { '200': { description: Shop updated, content: { application/json: { schema: { $ref: '#/components/schemas/ShopEnvelope' } } } } }
    delete:
      tags: [Merchants]
      summary: Delete a merchant shop
      operationId: deleteShop
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Shop deleted } }
...
  /api/v1/customers:
    get:
      tags: [Customers]
      summary: List customers visible to the authenticated membership
      operationId: listCustomers
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: Tenant- and shop-scoped customers, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } }
  /api/v1/customers/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Customers]
      summary: Get a customer
      operationId: getCustomer
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Customer, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    patch:
      tags: [Customers]
      summary: Update a customer as the merchant owner
      operationId: updateCustomer
      security: [{ bearerAuth: [] }]
      description: Requires the system `owner` or `merchant` role. Staff and manager memberships receive `FORBIDDEN` even if they otherwise have tenant write permission.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [display_name, customer_type], properties: { display_name: { type: string }, customer_type: { type: string, enum: [RETAIL, WHOLESALE, GUEST] }, email: { type: string, format: email, nullable: true }, phone: { type: string, nullable: true }, loyalty_identifier: { type: string, nullable: true } } } } } }
      responses: { '200': { description: Customer updated, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /api/v1/pos/shops/{shopId}/terminals:
    parameters: [{ $ref: '#/components/parameters/ShopID' }]
    get:
      tags: [Catalog]
      summary: List POS terminals for a shop
      operationId: listPOSTerminals
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: Shop terminals, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalCollectionEnvelope' } } } } }
  /api/v1/pos/terminals:
    post:
      tags: [Catalog]
      summary: Create a POS terminal for a shop
      operationId: createPOSTerminal
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalRequest' } } } }
      responses: { '201': { description: POS terminal created, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalEnvelope' } } } } }
  /api/v1/pos/terminals/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Catalog]
      summary: Get a POS terminal
      operationId: getPOSTerminal
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: POS terminal, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalEnvelope' } } } } }
    patch:
      tags: [Catalog]
      summary: Update a POS terminal
      operationId: updatePOSTerminal
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalRequest' } } } }
      responses: { '200': { description: POS terminal updated, content: { application/json: { schema: { $ref: '#/components/schemas/TerminalEnvelope' } } } } }
    delete:
      tags: [Catalog]
      summary: Delete a POS terminal
      operationId: deletePOSTerminal
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: POS terminal deleted } }
  /api/v1/pos/shops/{shopId}/sessions:
    parameters: [{ $ref: '#/components/parameters/ShopID' }]
    get:
      tags: [Catalog]
      summary: List POS sessions for a shop
      operationId: listPOSSessions
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Query' }, { $ref: '#/components/parameters/Filter' }, { $ref: '#/components/parameters/PageIndex' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: Shop POS sessions, content: { application/json: { schema: { $ref: '#/components/schemas/CollectionEnvelope' } } } } }
  /api/v1/pos/sessions:
    post:
      tags: [Catalog]
      summary: Open a POS session
      operationId: createPOSSession
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shop_id]
              properties:
                shop_id: { type: string, format: uuid }
                terminal_id: { type: string, format: uuid }
                membership_id: { type: string, format: uuid }
                opening_cash: { type: string }
      responses: { '201': { description: POS session opened, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } }
  /api/v1/pos/sessions/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Catalog]
      summary: Get a POS session
      operationId: getPOSSession
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: POS session, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } }
    patch:
      tags: [Catalog]
      summary: Close or update a POS session
      operationId: updatePOSSession
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [OPEN, CLOSED, ABANDONED] }
                expected_cash: { type: string }
                counted_cash: { type: string }
      responses: { '200': { description: POS session updated, content: { application/json: { schema: { $ref: '#/components/schemas/ResourceEnvelope' } } } } }
    delete:
      tags: [Catalog]
      summary: Delete a POS session
      operationId: deletePOSSession
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: POS session deleted } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send the access token returned by login or refresh as
        Authorization: Bearer <access-token>. Refresh tokens are rotated by
        POST /api/v1/auth/refresh.
  parameters:
...
    Query:
      name: query
      in: query
      required: false
      description: Case-insensitive text search over the endpoint's searchable fields.
      schema: { type: string }
...
    PageIndex:
      name: page_index
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    BadRequest:
      description: The request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              $ref: '#/components/examples/ValidationError'
...
    Forbidden:
      description: The authenticated principal lacks the required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist in the authorized tenant context.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ResourceEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: object
          description: Concrete endpoint resource payload.
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/ResourceMeta'
      additionalProperties: false
    CollectionEnvelope:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          description: Concrete endpoint resource collection.
          items:
            type: object
            additionalProperties: true
        meta:
          $ref: '#/components/schemas/CollectionMeta'
      additionalProperties: false
...
    RefundEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Refund' }
        meta: { type: object }
...